Removed require_once for Seeder
Implemented a more traditional "CodeIgniter" approach for the database Seeder. First, I removed the require_once to the Seeder file and the $root variable, the moved to the Seeder.php to libraries/Database/Seeder.php. Next in the Tools controller, I loaded the library normally using the Loader class. Also created the seeder object to access the newly moved seeder library.

Does not include tests since this repo is not built for PHPUnit. Works on a local XAMPP server running PHP 7.3.5